Fixed bug with java export.
git-svn-id: http://firstforge.wpi.edu/svn/repos/robotbuilder/trunk@246 7d5bc05a-d525-4a0d-8ff2-4902128d4f3b

+++ test/robotbuilder/JavaExportTest.java
@@ -33,12 +33,14 @@ public class JavaExportTest {
         File robotmap = new File("test-resources/RobotBuilderTestProject/src/robotcode/RobotMap.java");
         File oi = new File("test-resources/RobotBuilderTestProject/src/robotcode/OI.java");
         
-        for (File child : commands.listFiles()) { child.delete(); }
+        if (commands.listFiles() != null)
+            for (File child : commands.listFiles()) { child.delete(); }
         commands.delete();
         assertFalse(commands.exists());
         commands.mkdir();
 
-        for (File child : subsystems.listFiles()) { child.delete(); }
+        if (subsystems.listFiles() != null)
+            for (File child : subsystems.listFiles()) { child.delete(); }
         subsystems.delete();
         assertFalse(subsystems.exists());
         subsystems.mkdir();
